Add shared key generation
This function returns base58 encoded shared key from base58 encoded А private & B public keys for enhancement of the application as described in: https://github.com/wavesplatform/curve25519-js
The result should not be used directly as a key, but should be processed with a one-way function (e.g. HSalsa20 as in NaCl, or any secure cryptographic hash function, such as SHA-256, or key derivation function, such as HKDF).